disease-api/requirements.txt
flask==3.0.0
flask-cors==4.0.0
tensorflow-cpu==2.15.0
Pillow>=10.0.0
numpy>=1.24.0,<2.0.0
gdown==5.1.0
gunicorn==21.2.0
